Visual Studio 2013 で以下のエラーになるので修正
error C2668: 'round' : オーバーロード関数の呼び出しを解決することができません。(新機能 ; ヘルプを参照)
C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\include\math.h(1241): 'long double round(long double) throw()' の可能性があります。
C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\include\math.h(1125): または 'float round(float) throw()'
C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\include\math.h(516): または 'double round(double)'

Visual Studio 2013 から math.h に round() が含まれるようになったため
https://devblogs.microsoft.com/cppblog/c99-library-support-in-visual-studio-2013/

git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@8104 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/teraterm/addsetting.cpp
@@ -948,14 +948,14 @@ void CVisualPropPageDlg::OnOK()
 	// (1)
 	GetDlgItemTextA(IDC_ALPHA_BLEND_ACTIVE, buf, sizeof(buf));
 	if (isdigit(buf[0])) {
-		int i = (int)round((255 * atoi(buf) / 100));
+		int i = (int)round((double)(255 * atoi(buf) / 100));
 		ts.AlphaBlendActive =
 			(i < 0) ? 0 :
 			(i > 255) ? 255 : i;
 	}
 	GetDlgItemTextA(IDC_ALPHA_BLEND_INACTIVE, buf, sizeof(buf));
 	if (isdigit(buf[0])) {
-		int i = (int)round((255 * atoi(buf) / 100));
+		int i = (int)round((double)(255 * atoi(buf) / 100));
 		ts.AlphaBlendInactive = 
 			(i < 0) ? 0 :
 			(i > 255) ? 255 : i;
